Added AddSessionCommand
