+++ src/ClinicApp.Application/QueryServices/IDoctorQueryService.cs
@@ -4,4 +4,5 @@ namespace ClinicApp.Application.QueryServices;
 public interface IDoctorQueryService
 {
     Task<DoctorWithSessionsDTO?> GetDoctorWithSessions(Guid doctorid);
+    Task<DoctorWithSessionsDTO?> GetDoctorWithSessionsForDay(Guid doctorid, DateTime day);
 }
